JIRA:LV3-75[new feature] created pmtct-hts form

+++ src/main/java/org/lamisplus/modules/pmtct/domain/dto/PmtctVisitRequestDto.java
@@ -29,7 +29,7 @@ public class PmtctVisitRequestDto implements Serializable {
     private String timeOfViralLoad;
     private LocalDate dateOfViralLoad;
     private Integer gaOfViralLoad;
-    private Integer resultOfViralLoad;
+    private Long resultOfViralLoad;
     private String dsd;
     private String dsdOption;
     private String dsdModel;
