+++ src/main/java/org/lamisplus/modules/pmtct/domain/dto/PmtctVisitResponseDto.java
@@ -25,7 +25,7 @@ public class PmtctVisitResponseDto implements Serializable {
     private String timeOfViralLoad;
     private LocalDate dateOfViralLoad;
     private Integer gaOfViralLoad;
-    private Integer resultOfViralLoad;
+    private Long resultOfViralLoad;
     private String dsd;
     private String dsdOption;
     private String dsdModel;
